Added security for: CalendarTask, Category, TodoTask. Schedule added on creation of new user. CalendarTask (change from date to timestamp)

+++ ajdu_restful_api/src/main/java/ajdu_restful_api/controller/CategoryRestController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,7 +17,7 @@ import ajdu_restful_api.model.Category;
 import ajdu_restful_api.service.CategoryService;
 
 @RestController
-public class CategoryRestController {
+public class CategoryRestController extends AuthenticatedRestController {
 
 	@Autowired
 	private CategoryService categoryService;
@@ -27,13 +28,15 @@ public class CategoryRestController {
 	}
 	
 	@RequestMapping(value="/categories", method=RequestMethod.POST,consumes=MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Category> addCategory(@RequestBody Category cat) {
-		if(categoryService.findCategoryByName(cat.getName()) == null) {
-			categoryService.saveCategory(cat);
-			return new ResponseEntity<Category>(cat, HttpStatus.CREATED);
-		} else {
-			return new ResponseEntity<Category>(cat,HttpStatus.CONFLICT);
-		}
+	public ResponseEntity<Category> addCategory(@RequestBody Category cat, Authentication auth) {
+		if(isAdmin(auth)) {
+			if(categoryService.findCategoryByName(cat.getName()) == null) {
+				categoryService.saveCategory(cat);
+				return new ResponseEntity<Category>(cat, HttpStatus.CREATED);
+			} else {
+				return new ResponseEntity<Category>(cat,HttpStatus.CONFLICT);
+			}
+		} else return new ResponseEntity<Category>(HttpStatus.FORBIDDEN);
 	}
 	
 	@RequestMapping(value="/categories/{id}",method=RequestMethod.GET,produces=MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -45,23 +48,27 @@ public class CategoryRestController {
 	
 	
 	@RequestMapping(value="/categories/{id}",method=RequestMethod.DELETE,produces=MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Category> deleteCategory(@PathVariable int id) {
-		Category c = categoryService.findOneCategory(id); 
-		if(c != null) {
-			categoryService.deleteCategory(id);
-			return new ResponseEntity<Category>(HttpStatus.OK);
-		}
-		return new ResponseEntity<Category>(HttpStatus.NOT_FOUND);
+	public ResponseEntity<Category> deleteCategory(@PathVariable int id, Authentication auth) {
+		if(isAdmin(auth)) {
+			Category c = categoryService.findOneCategory(id); 
+			if(c != null) {
+				categoryService.deleteCategory(id);
+				return new ResponseEntity<Category>(HttpStatus.OK);
+			}
+			return new ResponseEntity<Category>(HttpStatus.NOT_FOUND);
+		} else return new ResponseEntity<Category>(HttpStatus.FORBIDDEN);
 	}
 	
 	@RequestMapping(value="/categories/{id}",method=RequestMethod.PUT,produces=MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Category> updateCategory(@PathVariable int id, @RequestBody Category category) {
-		Category c = categoryService.findOneCategory(id); 
-		if(c != null) {
-			c.setName(category.getName());
-			return new ResponseEntity<Category>(c,HttpStatus.OK);
-		}
-		return new ResponseEntity<Category>(HttpStatus.NOT_FOUND);
+	public ResponseEntity<Category> updateCategory(@PathVariable int id, @RequestBody Category category, Authentication auth) {
+		if(isAdmin(auth)) {
+			Category c = categoryService.findOneCategory(id); 
+			if(c != null) {
+				c.setName(category.getName());
+				return new ResponseEntity<Category>(c,HttpStatus.OK);
+			}
+			return new ResponseEntity<Category>(HttpStatus.NOT_FOUND);
+		} else return new ResponseEntity<Category>(HttpStatus.FORBIDDEN);
 	}
 	
 }
